// src/mvt/types.ts
export declare type ColorSpecification = string;
export declare type FormattedSpecification = string;
export declare type ResolvedImageSpecification = string;
export declare type PromoteIdSpecification =
  | {
      [_: string]: string;
    }
  | string;
export declare type FilterSpecificationInputType = string | number | boolean;
export declare type FilterSpecification =
  | ['at', number, (number | string)[]]
  | ['get', string, Record<string, unknown>?]
  | ['has', string, Record<string, unknown>?]
  | [
      'in',
      ...FilterSpecificationInputType[],
      FilterSpecificationInputType | FilterSpecificationInputType[],
    ]
  | [
      'index-of',
      FilterSpecificationInputType,
      FilterSpecificationInputType | FilterSpecificationInputType[],
    ]
  | ['length', string | string[]]
  | ['slice', string | string[], number]
  | ['!', FilterSpecification]
  | ['!=', string | FilterSpecification, FilterSpecificationInputType]
  | ['<', string | FilterSpecification, FilterSpecificationInputType]
  | ['<=', string | FilterSpecification, FilterSpecificationInputType]
  | ['==', string | FilterSpecification, FilterSpecificationInputType]
  | ['>', string | FilterSpecification, FilterSpecificationInputType]
  | ['>=', string | FilterSpecification, FilterSpecificationInputType]
  | ['all', ...FilterSpecification[], FilterSpecificationInputType]
  | ['any', ...FilterSpecification[], FilterSpecificationInputType]
  | ['case', ...FilterSpecification[], FilterSpecificationInputType]
  | ['coalesce', ...FilterSpecification[], FilterSpecificationInputType]
  | ['match', ...FilterSpecification[], FilterSpecificationInputType]
  | ['within', ...FilterSpecification[], FilterSpecificationInputType]
  | ['!in', ...FilterSpecification[], FilterSpecificationInputType]
  | ['!has', ...FilterSpecification[], FilterSpecificationInputType]
  | ['none', ...FilterSpecification[], FilterSpecificationInputType]
  | Array<string | FilterSpecification>;
export declare type TransitionSpecification = {
  duration?: number;
  delay?: number;
};
export declare type CameraFunctionSpecification<T> =
  | {
      type: 'exponential';
      stops: Array<[number, T]>;
    }
  | {
      type: 'interval';
      stops: Array<[number, T]>;
    };
export declare type SourceFunctionSpecification<T> =
  | {
      type: 'exponential';
      stops: Array<[number, T]>;
      property: string;
      default?: T;
    }
  | {
      type: 'interval';
      stops: Array<[number, T]>;
      property: string;
      default?: T;
    }
  | {
      type: 'categorical';
      stops: Array<[string | number | boolean, T]>;
      property: string;
      default?: T;
    }
  | {
      type: 'identity';
      property: string;
      default?: T;
    };
export declare type CompositeFunctionSpecification<T> =
  | {
      type: 'exponential';
      stops: Array<
        [
          {
            zoom: number;
            value: number;
          },
          T,
        ]
      >;
      property: string;
      default?: T;
    }
  | {
      type: 'interval';
      stops: Array<
        [
          {
            zoom: number;
            value: number;
          },
          T,
        ]
      >;
      property: string;
      default?: T;
    }
  | {
      type: 'categorical';
      stops: Array<
        [
          {
            zoom: number;
            value: string | number | boolean;
          },
          T,
        ]
      >;
      property: string;
      default?: T;
    };
export declare type ExpressionSpecificationArray = Array<unknown>;
export declare type PropertyValueSpecification<T> =
  | T
  | CameraFunctionSpecification<T>
  | ExpressionSpecificationArray;
export declare type DataDrivenPropertyValueSpecification<T> =
  | T
  | CameraFunctionSpecification<T>
  | SourceFunctionSpecification<T>
  | CompositeFunctionSpecification<T>
  | ExpressionSpecificationArray;
// ...
export declare type FillLayerSpecification = {
  id: string;
  type: 'fill';
  metadata?: unknown;
  source: string;
  'source-layer'?: string;
  minzoom?: number;
  maxzoom?: number;
  filter?: FilterSpecification;
  layout?: {
    'fill-sort-key'?: DataDrivenPropertyValueSpecification<number>;
    visibility?: 'visible' | 'none';
  };
  paint?: {
    'fill-antialias'?: PropertyValueSpecification<boolean>;
    'fill-opacity'?: DataDrivenPropertyValueSpecification<number>;
    'fill-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>;
    'fill-outline-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>;
    'fill-translate'?: PropertyValueSpecification<[number, number]>;
    'fill-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>;
    'fill-pattern'?: DataDrivenPropertyValueSpecification<ResolvedImageSpecification>;
  };
};
export declare type LineLayerSpecification = {
  id: string;
  type: 'line';
  metadata?: unknown;
  source: string;
  'source-layer'?: string;
  minzoom?: number;
  maxzoom?: number;
  filter?: FilterSpecification;
  layout?: {
    'line-cap'?: PropertyValueSpecification<'butt' | 'round' | 'square'>;
    'line-join'?: DataDrivenPropertyValueSpecification<
      'bevel' | 'round' | 'miter'
    >;
    'line-miter-limit'?: PropertyValueSpecification<number>;
    'line-round-limit'?: PropertyValueSpecification<number>;
    'line-sort-key'?: DataDrivenPropertyValueSpecification<number>;
    visibility?: 'visible' | 'none';
  };
  paint?: {
    'line-opacity'?: DataDrivenPropertyValueSpecification<number>;
    'line-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>;
    'line-translate'?: PropertyValueSpecification<[number, number]>;
    'line-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>;
    'line-width'?: DataDrivenPropertyValueSpecification<number>;
    'line-gap-width'?: DataDrivenPropertyValueSpecification<number>;
    'line-offset'?: DataDrivenPropertyValueSpecification<number>;
    'line-blur'?: DataDrivenPropertyValueSpecification<number>;
    'line-dasharray'?: PropertyValueSpecification<Array<number>>;
    'line-pattern'?: DataDrivenPropertyValueSpecification<ResolvedImageSpecification>;
    'line-gradient'?: ExpressionSpecificationArray;
  };
};
// ...
export declare type CircleLayerSpecification = {
  id: string;
  type: 'circle';
  metadata?: unknown;
  source: string;
  'source-layer'?: string;
  minzoom?: number;
  maxzoom?: number;
  filter?: FilterSpecification;
  layout?: {
    'circle-sort-key'?: DataDrivenPropertyValueSpecification<number>;
    visibility?: 'visible' | 'none';
  };
  paint?: {
    'circle-radius'?: DataDrivenPropertyValueSpecification<number>;
    'circle-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>;
    'circle-blur'?: DataDrivenPropertyValueSpecification<number>;
    'circle-opacity'?: DataDrivenPropertyValueSpecification<number>;
    'circle-translate'?: PropertyValueSpecification<[number, number]>;
    'circle-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>;
    'circle-pitch-scale'?: PropertyValueSpecification<'map' | 'viewport'>;
    'circle-pitch-alignment'?: PropertyValueSpecification<'map' | 'viewport'>;
    'circle-stroke-width'?: DataDrivenPropertyValueSpecification<number>;
    'circle-stroke-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>;
    'circle-stroke-opacity'?: DataDrivenPropertyValueSpecification<number>;
  };
};